sdhci: Avoid tuning on ADTC commands.
BZ: 161050

The sdio host controller driver perform a tuning on all kind of CMD52
even during CMD52 in write mode. We noticed that Broadcom wifi card
does not execute the tuning command while sleeping. As a result, the
host controller driver fails on the tuning commmand and switch back to
fixed sampling rate.

This issue is observed only in SDR104 mode. To fix the issue, the host
controller driver shall avoid executing the tuning command on ADTC
commands.

Change-Id: I998d1f28da9bc38aadf8562da3e6404f274489ed
Signed-off-by: Aymen Zayet <aymen.zayet@intel.com>