drm/i915: set the brightness logic max level to 100 for VLV2 MIPI
BZ: 162554

Currently for VLV2 mipi, the max level of brightness is 0xFF, which
is impacting the system resuming/suspending time indeed.

Because the max level value is bigger, the brightness from low to high
will take more steps to achrieve.

Compared to CTP platform, it also set the logic max level to 100 although
the max level value is 0xff.

So here optimizing it for some KPI target.

Change-Id: I5ac4894ec36cd8d39fc1952b18d5bdf6a18df2d9
Signed-off-by: Liu, Chuansheng <chuansheng.liu@intel.com>