remove pm get for acm read
BZ: 163524

In ACM and NCM, pm get and put are added to their
read callback. As their parents is the same USB
device. This will lead to concurrent issue like
spin lock bad magic in pm put. So remove pm get
and put for ACM. Keep this change only for NCM.
NCM need this to support 3G FTP use cases.

Change-Id: I38bb8d08895ba39ec591d4bdae29c1371df3aa65
Signed-off-by: feng wang <feng.a.wang@intel.com>