EM/Charger: Modifying the charger health logic
BZ: 160468

Problem : In the case of byt ffd8 pr0 STAT_D register is
not used for detecting the charger type. which results the charger
health to report Unknown always and resulting the battery status
remain Not charging

Solution :In case of pr0 the detection logic is in fsa
so reporting charger health only by checking the charger is enabled.

Change-Id: I5769d2edb032edac4b062f30f7671a81b66a26ca
Signed-off-by: Gaurav <gauravx.thakur@intel.com>