EM/Charger: Fix the undervoltage status at higher capacity
BZ: 162385

When the capacity reaches 80% or above, the charging status
toggles between charging and not-charging, with charging LED
toggling due to charger goes undervoltage condition and recovers
after that.
This is root caused to switching frequency and
turbo charge enable. This patch sets the switching frequency
and turbo charge enable to the charger NVM settings. This patch
use the USB-OK stat instead of charging stat for STAT pin output.

Change-Id: I5c9a8ca37eb5b510956152564724ae2656e130c4
Signed-off-by: R, Kannappan <r.kannappan@intel.com>