gpio-virtual: Add Virtual GPIO controller driver
BZ: 160650

Due to GPIO pin BUG, some HW devices can't wake-up using
the designed flow through GPIO controller. For this devices
the alternative route is to contact PMC. If the device is PME
enable, PMC will forward SCI to IA, which will land on
IRQ 9.

This patch add driver for Virtual GPIO controller which
occupies IRQ 9. When an IRQ will land on IRQ 9, the IRQ
routine will clear the STS bit to clear the SCI.

Change-Id: I1037dd2d7e0789933db562302a711d7404bec37f
Signed-off-by: Souvik Kumar Chakravarty <souvik.k.chakravarty@intel.com>