Audio: SST: Fix sst_state in shutdown and control
BZ: 168431

When the driver is getting unloaded, use SHUTDOWN state
instead of RESET, so that further handling of any other call
is not done.
Also in the device/cdev controls, commands for pause/resume et
al needs to be handled only on an open stream where Fw state
is in RUNNING so change it accordingly.

Change-Id: Ib397e681eaa81f43e4fbc6636e0b7c8d1684ad0e
Signed-off-by: Mythri P K <mythri.p.k@intel.com>