Audio: SST: DMA sglist length check
BZ: 160919

When stitched tool generated FW binary(bin) is used instead of elf,
there is an additional SST_CUSTOM_INFO (sst ram type) block. Driver
needs to ignore this and not add it for DMA copy when generating
the sglist for I/DRAM and DDR block. This should be taken care of
when calculating the sg_list length as well, otherwise if there
is a mismatch between sg_list length and block data, there is
chance of writing into bad address resulting in unrecoverable crash.

Change-Id: I9e67c0ff47a6c3293faa51c442cb89f1cde6c09a
Signed-off-by: Mythri P K <mythri.p.k@intel.com>
Signed-off-by: mallikarjuna <mallikarjunareddyx.amireddy@intel.com>