NPK: Initial implementation of the North Peak driver
BZ: 171047

The North Peak (NPK) hardware is a set of functional blocks whose
primary purpose is to enable external customers, as well as internal
validation teams, to debug Intel Architecture (IA) based systems using
off-the-shelf third-party debug tools in addition to internal
tools. The North Peak hardware is implemented as a PCI device.

The North Peak SDK is a software package for trace collection and
visualization for the North Peak trace hub. This package includes for
example:
 - a software trace instrumentation API (SVENTX) usable in different
   software environments (user-space, kernel, firmware...)
 - an API and tool to configure the North Peak hardware
 - a driver to communicate with the North Peak device

This change is the initial implementation of the North Peak kernel driver.
Its main functionalities are:
 - SW master/channel management
 - read/write configuration registers
 - trace buffer management

Change-Id: I06ce5ede2d6c2eb638b951aa304e93f778487147
Signed-off-by: Yann Fouassier <yann.fouassier@intel.com>