drm/i915: Native sync stale sync point fix
BZ: 172302

The driver uses the HW seqno counter as the monotonic timeline counter.
When the ring wraps the timeline is reset back to 0, however the sync
framework may request that a sync point is duplicated, even after it is
signalled.

This patch adds an extra field to the timeline and to sync points.
When a sync point is created it is stamped with the timelines current cycle
count. When the timeline is reset the "cycle" count is incremented.
A sync point belongs to a particular cycle so if its status is queried
and the cycle value of the sync point does not match the cycle value of the
timeline then the point must have been passed and it is marked as "signalled".

In theory the cycle count could wrap and if a sync point was kept around
long enough then the timeline cycle could become equal to the sync point cycle,
but in practice the seqno count would have to wrap 2^32 times.

This patch also removes a call to sync_pt_free in the error handling case
of i915_sync_finish_request as this function is not the owner of pt.
Under certain failure conditions the same sync point could be free'd
twice resulting in a NULL pointer dereference.

Change-Id: I82773d77ae48c3a0c7f2e56b377f65d19c12c0f0
Signed-off-by: jzielins <jan.zielinski@intel.com>