Audio: SST: Remove csr lock and use sst_lock instead
BZ: 173533

Whenever CSR register is accessed to change the LPE state, CSR lock is used.
Scenarios when LPE status is changed is either during loading of fw/lib or
in recovery, fw/lib loading scenario is already handled with sst_lock, and
currently CSR is written without a lock in recovery, so cover the scenario
with sst_lock and remove csr_lock.

Change-Id: Idf7876fd2dde5d796895aa81bb867f35b2f440be
Signed-off-by: Mythri P K <mythri.p.k@intel.com>