Audio: enable/disable SSP port when stream is active/inactive
BZ: 165559

In case when playback and capture stream is active, if one of the
stream is closed, the ssp port gets disable as the SSP configuration
is done separately for each endpoints.

This patch adds :
 - added codec-codec dai link to use the BE CPU dai for SSP port
 - added new codec-codec dai link for modem to use BE CPU dai with
dsp_loopback enabled
 - BE CPU Dai for SSP ports and dai_ops for startup and shutdown to
disable and enable SSP port.
 - enabling/disabling the SSP port when stream is activated
and disabled when all stream are inactive
 - Modified the machine dapm route map to connect the backend
endpoints.

Change-Id: I5b671da87daa46ba2c892c1a078b647b2c22f405
Signed-off-by: Jeeja KP <jeeja.kp@intel.com>
Signed-off-by: Omair Mohammed Abdullah <omair.m.abdullah@intel.com>