pci: Manage kernel panic mode
BZ: 172526

If kernel panic occurs, interrupts are disabled.
To save kernel logs into emmc, the panic handler calls
a power up PCI api to move sdhci block to D0 state.
If the sdhci block is in D3 state, the pci API calls msleep
that leads to re-enable the interrupts. To avoid double panic
due to the reeanbling of the interrutps, the panic kernel mode
is checked and udelay is called instead of usleep

udelay does not re-enable interrupts whereas usleep does

Change-Id: I767df5669d5ed1133e3288c38749a766159e3d0f
Signed-off-by: Christophe Guerard <christophe.guerard@intel.com>