mmc: fix rpmb switch error.
BZ: 162995

If emmc works as HS200 on rpmb, we can see failure when
it switches back from rpmb to user area. Root cause is
sampling clock is cleared during d0i3 and tuning is not
allowed on rpmb per spec. So, this patch downgrade speed
mode to DDR50 due to this limitation.

Change-Id: Idf6ab598efffd135fd6d81666a1f7a9155de1f7f
Signed-off-by: yisun1 <yi.y.sun@intel.com>