kct: Add filelist support as attachment
BZ: 174712

In order to be able to copy files during a tp2e event, we need to
attach the filename of these files to the KCT event. Their names are
concatenated in a single string separated by semicolons.
The kct event structure is updated accordingly, and the macros are
adapted to accept one more argument.

Change-Id: Iaa3c6eebedb819b15c279ecda9c72a762ca6d18f
Signed-off-by: Flavien Boutoille <flavienx.boutoille@intel.com>