Audio: Fix algo params size in lpe_mixer_algo_params
BZ: 173524

KW issue in 64 bit kernel is because, the size to algo params is assigned as
sizeof algo_params which is 16 bytes. and the params pointer is assigned to
mixer_params which is of size 12 bytes. So during memcpy in sst_send_algo_param,
it may try to access indexes 12~15 which is invalid.

Change-Id: Ibca4acf9549140d4ce90809447330057f84d6c91
Signed-off-by: Subhransu S. Prusty <subhransu.s.prusty@intel.com>
Signed-off-by: mallikarjuna <mallikarjunareddyx.amireddy@intel.com>