Audio: SST: Replace stream_lock with sst_lock
BZ: 173533

Currently stream lock is used in open and release ctrl, but open
is already protected for concurrency with sst_lock, and release
has only runtime_put where all the operations are done in atomic context
so we don't really need a lock there.
Other instance where it is used is in clearing a stream in case of
invalid stream-id, where again sst_lock can be used.

Change-Id: I52f8ecb58411459a1d0e34bff85f86a8d2ff1f8f
Signed-off-by: Mythri P K <mythri.p.k@intel.com>