Audio: SST: Replace mixer_ctrl_lock with sst_lock
BZ: 173533

mixer_ctrl_lock is used when setting the algo params, currently
used in byt and ctp. This is used in combination with sst_check_device
which in turn uses sst_lock for concurrency check, same lock
can be used for mixer control setting.So replace mixer_ctrl_lock
with sst_lock.

Change-Id: Ib2c089622d32a14af8184b07648c0fb85b27ecbb
Signed-off-by: Mythri P K <mythri.p.k@intel.com>