drm/i915: Scratch page optimization for blanking buffer
BZ: 186003

When HWC is trying to flip a blanking buffer on primary plane,
we setup all the GTT entries of that blanking buffer with scratch
page (which is already zeroed out). This saves the time in allocation
of backing physical space for the blanking buffer and flushing of CPU
cache.

Change-Id: I9682ad8646945dc06e89d37bda36434d457ce04b
Signed-off-by: Akash Goel <akash.goel@intel.com>