[PORT to DPCM][REVERTME]audio:Do not use VLEX bypass mode for audio
BZ: 176391

During BIAS_OFF we configure the PMIC to use VFLEX
bypass mode. But this is not working in Shadycove PMIC.

This workaround forces to keep the vflex audio rail to
keep on even during audio idle for MOFD platform.

This should be reverted once the issue is root caused.

Original-Change-Id: I01349af5b9c5c279a9c2f54e316f1b7562ca971b
Original-commit 6864fae052824f935e97f0d37827bf0cfe626253

Change-Id: I6aaa23204197e27f344c7837c98d8a86c7c56ba2
Signed-off-by: Desetty ManikanthaX <manikanthax.desetty@intel.com>