gpio-langwell: fix the wrong GIMR operation when requesting GPIO
BZ: 175070

When we request a GPIO pin, lnw_gpio_request() will be called
to set GAFR of requested pin to 0. However, in moorefield platform,
there is no GAFR, and the GAFR address calculated by gpio driver
is GIMR in moorefield.

We skip the wrong GAFR/GIMR operation when requesting GPIO.

Change-Id: Ida778acbb68da0e5a1b5b2842d5dc46e159f0d0c
Signed-off-by: Li, Ning <ning.li@intel.com>