drm/i915: Parse EDID probed modes for DRRS support
BZ:175037

This patch finds out the lowest refresh rate supported for a given
resolution, generally the fixed_mode for LFP, based on the logic
implemented in find_panel_downclock. It also checks the VBT fields
to see if panel supports seamless DRRS or not. Based on above data
it marks whether eDP panel supports seamless DRRS or not.
This information is needed for supporting seamless DRRS switch for
certain power saving usecases. This patch is tested by enabling the
DRM logs and user should see whether Seamless DRRS is supported or not

Change-Id: I541a1b753e9d724a04d0aaf2be0831bb06d580f9
Signed-off-by: Akashdeep Sharma <akashdeep.sharma@intel.com>
Signed-off-by: Uma Shankar <uma.shankar@intel.com>
Signed-off-by: Pradeep Bhat <pradeep.bhat@intel.com>
Signed-off-by: Vandana Kannan <vandana.kannan@intel.com>