Revert "ASoC: wm_adsp: Bail out earlier if a control already exists"
This reverts commit 9e9d1202ae9cdc67a02db8b5a388e4a9c743ea03.

This commit causes the later sections of larger controls not to be
enabled when the firmware is loaded for a second time. The control
enabling and control creation are currently rather heavily intertwined
in the code and seperating them will require more significant
refactoring.

As the original patch is a performance improvement rather than a bug fix
revert until the refactoring can be done.

Change-Id: If4fb3fa70d7547105f4f9b6cdbe75608bbe9128b
Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>