EM/THERMAL: Add emulation support to MRFLD thermal driver
BZ: 194468

This patch adds emulated temperature support to the
MRFLD thermal driver. This is enabled only when
CONFIG_THERMAL_EMULATION is defined. When enabled,
this creates virtual thermal zones under
/sys/class/thermal/ which can take (fake)temperature
as input from user/tools.

When the user reads 'temp' interface, the driver
requests the user space for the (calibrated)
temperature through UEvent. On a write from
user space into the 'emul_temp' sysfs interface,
the driver returns this temperature to the user;
as a response to the original read on the 'temp'
interface.

The user/tools are abstracted from all the
complexity since they still work with the
standard 'temp' interface and get the calibrated
temperature as a response.

Change-Id: I163dccc2faafdb8e8215c2b2e912a8913d45db1e
Signed-off-by: Ajay Thomas <ajay.thomas.david.rajamanickam@intel.com>