dwc3-intel-mrfl: Fix the UTMI phy can't working with PMIC A1 issue.
BZ: 177724

The USBPHYCTL.STB(bit0) is control the switcher inside of PMIC to
connect/disconnect the DP/DM between PMIC and SOC.

This bit can't working with PMIC A0. Whatever set or not, the DP/DM
always connected. But with PMIC A1, this issue be fixed. So need driver
to re-connect them after resume from D0i3cold.

Change-Id: I224b74a7b80d0569f5dd4de3f0599a6429688d38
Signed-off-by: Wang, Yu <yu.y.wang@intel.com>