[PORT FROM R44B]mmc: The sdio func device is added to the driver model after the card device.
BZ: 180260

This patch is ported from linux-mmc community, which is still in review process

This means the sdio func device will be suspend before the card device
and thus resumed after. The consequence are the mmc core don't
explicity need to protect itself from receiving sdio requests in
suspended state. Instead that can be handled from the sdio bus, which
is thus invokes the PM callbacks instead of old dummy function.

In the case were the sdio func driver don't implement the PM callbacks
the mmc core will in the early phase of system suspend, remove the
card from the driver model and thus power off it.

>Cc: Aaron Lu <aaron.lu@intel.com>
>Cc: NeilBrown <neilb@suse.de>
>Cc: Rafael J. Wysocki <rjw@rjwysocki.net>
>Signed-off-by: Ulf Hansson <ulf.hansson@linaro.org>

Change-Id: Ic75a57d9888834f1cfe899997ed710d45f07e3eb
Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>