audio: fix wrong typecast in ssp slot map
BZ: 178218

For the SSP slot map command to be sent after pipeline has been turned on, a
valid widget pointer is needed. Rework the SSP_SLOT_CTL so that the private data
has a widget pointer. The power status of this widget pointer is checked before
sending the SLOT_MAP command.

Change-Id: Id9e76f81133d506d7afaf25547977b28c8c54208
Signed-off-by: Omair Mohammed Abdullah <omair.m.abdullah@intel.com>