audio: send VB_START before enabling SSP
BZ: 178218

The SBA timer needs to be enabled at the start of any UC, and disabled at
the end of all UCs. Handle this by calling VB_START/IDLE in the BE
startup/shutdown. A reference counter is used to ensure that the timer is
turned off only when nothing is running.

Change-Id: I68bd7de2bda973bb07c70a4f05f04334d80c9b57
Signed-off-by: Omair Mohammed Abdullah <omair.m.abdullah@intel.com>