Camera: Add support for the FSA642 CSI lane mux
BZ: 181315

In some HW configurations, there is a CSI lane multiplexer switching
between the OV680 external ISP and the IMX132 front camera sensor.
Ensure that the mux is always set to route the IMX132 sensor signals
through, when the IMX132 sensor is powered on.

Please note that the variation on whether this mux exist in the HW
or not is done by looking at the GPIO list in the firmware. If a
GPIO with the correct name is found, then the mux is present and it
needs to be controlled.

Change-Id: I71360c88f9f6efa1e79a45ec347ec5c418b61154
Signed-off-by: Soren Friis <soren.friis@intel.com>