CPU ConCurrency (CC): measure CPU load by averaging running tasks
BZ: 180965

CC is a CPU semantic that measures the load of the CPU by averaging the
number of runnable tasks on it. Then PM module can leverage CC to shape the CPU
utilization, i.e., consolidating light workload to run on smallest number of
CPUs that are capable of running it.

Change-Id: Id68538127d130889e1e016562b88c5ee45a404f3
Signed-off-by: Yuyang Du <yuyang.du@intel.com>