imx132: Prevent multiple GPIO searches when the FSA642 mux is not present
BZ: 182345

The code in platform_imx132 that is used to search for the GPIO
that controls the FSA642 CSI bus mux, will keep searching for the
GPIO when the mux is not present.

Fix the logic so the GPIO search is done only once, both when the
mux is present and also when it is not present.

Fix also the initialization of the static variables so they cannot
be reset if someone calls imx132_platform_data after the initial
call during boot.

Change-Id: Ie46492c7a50bae117693afad91c21aa1221322ad
Signed-off-by: Soren Friis <soren.friis@intel.com>