vibra:Add a flag to indicate if vibra needs gpio enable
BZ: 173675

Most vibra HW implementations use a gpio to enable the vibra driver.
But some do not. Hence add a platform data flag to indicate if
vibra on the particular platform uses gpio based enable control.

This flag can be used to skip gpio operations if it is not applicable
on a given platform. This helps the re-use of certain vibra functions
involving gpio operations across platforms.

Change-Id: Iac2199b6d4dda2317379fbe0aae5baafd7627b14
Signed-off-by: Jayachandran B <jayachandran.b@intel.com>