smiapp: Add register diversion quirk
BZ: 183309

Add a quirk for diverting registers for on some sensors, even the standard
registers are not where they can be expected to be found. Add a quirk to
to help using such sensors.

smiapp_write_no_quirk() and smiapp_read_no_quirk() functions are provided
for the use of quirk implementations.

Change-Id: Ia036e375f3eca7d07dd2fe064ed70ef2244f979c
Signed-off-by: Sakari Ailus <sakari.ailus@linux.intel.com>