smiapp-pll: Add quirk for op clk divisor == bits per pixel / 2
BZ: 183309

For some sensors in some configurations the effective value of op clk div is
bits per pixel divided by two. The output clock is correctly calculated
whereas some of the rest of the clock tree uses higher clocks than
calculated. This also limits the bpp to even values if the number of lanes
is four.

Change-Id: I37417b0fbc468f894afae8da94af8ed3cc580b30
Signed-off-by: Sakari Ailus <sakari.ailus@linux.intel.com>