dwc-host-intel: Change USB compliance test sysfs interface.
BZ: 194423

The compliance test sysfs can only show up when attach Micro-A cable
before, so it is inconvenient for tester to handle sysfs with debug
board. The sysfs will show up when dwc driver probe now, and tester
can handle it with USB interface first and then attach Micro-A cable
to run test.

Change-Id: I0bb94d00f5b837bdd1bc184f7c480b5b1df7c530
Signed-off-by: Li,Guanglei <guangleix.li@intel.com>
Signed-off-by: Wang, Yu <yu.y.wang@intel.com>