[MOFD]PM: Fix code for C6 check in C-state substate calculation
BZ: 146915

The "if ((cstate + 1) < 6)" comparison in the code assume 6 to represent C6.
But that’s not correct. “cstate” here is index in table and not C-x number.
We can’t hardcode the ‘6’ here b’coz the index will change if you add/remove
c-states

The patch fixes the issue by comparing against mwait_cstate instead of cstate
variable. mwait_cstate is 5 for C6 (hint=0x52). This is unlikely to change
for IA and removes the dependency on the table entries

Change-Id: I8364c442abc49cdda482a5e567399953dc91a1b5
Signed-off-by: Rushikesh Kadam <rushikesh.s.kadam@intel.com>