[REVERT ME] APIC: Add suspend/resume handling for TSC-deadline mode for CHT
BZ: 193429

This patch adds the support for suspend/resume handling
for TSC deadline mode for CHT.

This is a temperory hack until we figure out a way to
generalise the condition for all the tablet/phone platforms
for upstreaming purpose.This is currently looked upon. Until then this
fix the crash on CHT S3 cycle, we need to get this patch in.

Change-Id: I6ed36f9bcc8088733ea2118dfb30ae2ef0895f8c
Signed-off-by: Kumar Mahesh <mahesh.kumar.p@intel.com>