fsa642: Make the mux visible to other drivers
BZ: 184969

In some HW configurations there is an FSA642 CSI multiplexer that
controls whether the CSI lanes from the IMX132 sensor or the OV680
will get routed to the SoC.

Both drivers contained code to switch the mux in their favor but
unfortunately the search for the GPIO number could only be done once.
When the other driver searched, it would receive back an error.

Add a new platform file which controls the GPIO for the FSA642 mux
and call that from the IMX132 and OV680 platform drivers.

Change-Id: Ica4c182d3601c4d3b1757ff4c7d1ba71215dcb40
Signed-off-by: Soren Friis <soren.friis@intel.com>