[Re-Work]drm/i915: hack for pdf flicker issue
BZ: 190200

When in linear memory we often see flicker and the root cause is yet to
be analyzed. As an intermediate fix to this, the DDL has been set to '0'
when we are in linear memory type.

Note: This is just a quick fix/hack to unblock. Need to root cause the
issue and provide proper fix.

Change-Id: Ifbd45c3efe0683c0e499087c87b6cf44ca816f72
Signed-off-by: Arun R Murthy <arun.r.murthy@intel.com>