EM/THERMAL: MOFD: Add emulation support to thermal driver
BZ: 188343

[Implementation]
This patch adds emulated temperature support to the
MOFD thermal driver. This is enabled only when
CONFIG_THERMAL_EMULATION is defined. When enabled,
this creates virtual thermal zones under
/sys/class/thermal/ which can take (fake) temp.
as input from user/tools.
When the user reads 'temp' interface, the driver
requests the user space for the (calibrated)
temperature through UEvent. On a write from
user space into the 'emul_temp' sysfs interface,
the driver returns this temperature to the user;
as a response to the original read on the 'temp'
interface.
The user/tools are abstracted from all the
complexity since they still work with the
standard 'temp' interface and get the calibrated
temperature as a response.

Change-Id: I155350adc836c260e3032597505ed4217251f8f2
Signed-off-by: Sumeet Pawnikar <sumeet.r.pawnikar@intel.com>