eHCI driver: Enable use of SRAM
BZ: 180249

On 32-bit kernel, the eHCI driver uses the DMA coherent API to manage the SRAM.
This API is conditionned by the CONFIG_HAVE_GENERIC_DMA_COHERENT flag in the kernel.
The CONFIG_HAVE_GENERIC_DMA_COHERENT flag is not supported (so far) on 64-bit kernel,
and is not planned to be supported (see BZ188004).

With this patch, the eHCI driver manages the SRAM through its own memory allocator.

Change-Id: I2b078df67291e06cfc5e154904cef7810c95f441
Signed-off-by: Isabelle Leonardi <isabellex.leonardi@intel.com>