EM/THERMAL: Modify PMIC DIE T_CRIT setting
BZ: 181102

The register description for the PMIC DIE
critical temperature has changed between CRC
and CRC+ PMICs. The get/set_alert_temp
code in this driver works correctly for
CRC PMIC and not for CRC+ PMIC.

This patch addresses the issue by differentiating
the two different PMIC versions.

The ideal way is to use PMIC's 'vendor id -
product id' combination to achieve this; but
these registers are not mapped onto I2C space,
and hence not visible to the OS. Hence, use
CPU model id to differentiate them. This works
because we only have VLV platforms (using CRC)
and CHV platforms (using CRC+).

Change-Id: I0c3b55e5a219c3a88e8895931c16033baf0dbd22
Signed-off-by: Durgadoss R <durgadoss.r@intel.com>