Audio: soc-fw: Add the bytes_ext type kcontrol parser
BZ: 191883

Bytes ext uses the soc_byte_ext struct and hence cannot fit into
the regular parsing logic of mixer or enums.
Add support for bytes_ext type that will used for the like of algo_controls.
This patch also changes soc.h to include the list for bytes type
along with mixers and enum that is currently added.

Change-Id: I2f5f7740488dc850c625cd2688decd1bac64c23a
Signed-off-by: Mythri P K <mythri.p.k@intel.com>