[REVERTME][BYT CR]Switch between BT and built-in speaker/mic during VOIP call will cause Media service died
BZ: 191499

For Baytrail CR, we get CFG_LOx.FIFO_EMPTY about 120 loops.
So I have to increase max_loops(timeout) from 100(4.95ms) to 200(19.9ms).
If we try to disable DMA channel without waited FIFO_EMPTY,
channel cannot be disabled successfully and the whole DMA will be stuck.

This is a workaround from software, it should be a SoC fixing for
[BYT_CR_2.0][R4.4b-Stable][Audio_DMA]DMA sticks when disable channel before CFG_LOn.FIFO_EMPTY is true
https://vthsd.fm.intel.com/hsd/tablet_platform/sighting/default.aspx?sighting_id=5218671

Change-Id: I09d141a7c1a29bea77f52b077124f7385d0bed0f
Signed-off-by: Chen, Peter <peter.chen@intel.com>