input: simulate key release only for keys pressed during suspend entry.
BZ: 186514

Input framework simulates key release for every key detected as pressed
during suspend exit. This is wrong if the key was not actually pressed
during suspend entry. This patch ensures that the framework do not
simulate a key release for keys which were not actually pressed during
suspend entry.

Change-Id: I75ace29ce862e8d4e235abbf55aad4ac6cedb638
Signed-off-by: Catalin Popescu <catalin.popescu@intel.com>