drm/i915: Backlight fix for 720p scalar optimization
BZ: 190808

This change fixes the issue which happens due the HWC changes
(enabling panel fitter), which dynamically changes frame buffer
based on the content playing, since DPST calculations are based
on the resolution, we were seeing this issue.
This patch updates the threshold values for DPST with new resolution
values and updates the user mode with the change of resolution.

Change-Id: I996542ffdc42e89358ceba648a22da6ce86741a4
Signed-off-by: sgurram <sudheer.gurram@intel.com>