drm/i915: Fix to update latency values for sprite planes.
BZ: 195169

The drain latency values for sprite planes is updated in the DDL registers
when it is enabled for the first time. After any subsequent disable and
re-enable of the sprite plane, the drain latency value for sprite plane is
programmed to 0. Adding fix to update the correct drain latency..

Change-Id: I3f97491333c0c48f65f67da962080f29d95e2dd6
Signed-off-by: Gagandeep S Arora <gagandeep.s.arora@intel.com>