usb: dwc3: gadget: Fix scatter-gather list usage
BZ: 198582

When using scatter-gather lists in USB requests, the DWC3 interrupt
is supposed to be triggered after the last sg element (TRB).
This behavior is broken: the last sg element is not marked as the
last TRB because we check the USB request is the last of a list.
But preparing the TRBs moved the request from the list, hence this
checks never pass, so the "last" flag is not set, so the interrupt
is never triggered.

Change-Id: If6080aec93d573e0207205b13d0eab609d100a8b
Signed-off-by: Flavien Boutoille <flavienx.boutoille@intel.com>
Signed-off-by: Jeremy Rocher <jeremy.rocher@intel.com>