sep54: Support for 4G+ memories in 64bit system
BZ: 191755

There's only 32 bits reserved for cookie addresses in chaabi
descriptor messages. Previously this was bypassed by using
physical addresss. But this works only if the physical memory map is
in 32bit area. Added a static array for 64bit addresses and using
the index of the array as a cookie.

Change-Id: Iceec10bbe5713f0d3b2ead57f7242063e015702e
Signed-off-by: Santeri Salko <santerix.salko@intel.com>