dwc3: otg: cht: set timeout of cfg1 polling to 1 sec
BZ: 192785

Problem:
After plugging usb cable, when switching the mux to device controller,
after setting cfg0 with SW_IDPIN=1, Host Owned bit (bit 29) of cfg1 is
not cleared after a long time (about 500 ~ 600 ms)

Impact:
Currently, the timeout value in driver is 500 ms, which is not enough.
So dwc3 driver can set the run/stop bit before PHY is switched to dwc3
controller. And we'll see a Erratic Error Reported. But the enumeration
will be Okay.

Thus, make the timeout value as 1000 ms to guarantee PHY is switched
before run/stop bit is set.

Change-Id: I8ffdb6358d6b58baa91622cbbd08de54bafd9a24
Signed-off-by: Zhuang Jin Can <jin.can.zhuang@intel.com>