usb: dwc3: otg: extend ulpi write timeout to 10 ms
BZ: 194206

Although most of the time, ulpi write only requires about
1~5 us, it's oberserved that ulpi write sometimes could more
than 3 ms.

Before disconnecting the pullup, dwc3 needs to put the PHY in
non-driving mode via a ulpi write, otherwise, the PHY may hang.
Thus if the ulpi write timeouts due to not enough polling time,
we'll see PHY hang.

The same should be applied to ulpi read.

The patch also actives the basic otg log, so that if a ulpi
write fails, we can see it from aplog.

Change-Id: Ia111fbf37d375be527241eeb9550a4e9fb04ce02
Signed-off-by: Zhuang Jin Can <jin.can.zhuang@intel.com>