audio: vtsv: cache the VTSV enrollment data
BZ: 194007

Remove the dependency of sequence on userspace by caching the enrollment
parameters. The driver now caches the enrollment parameters (net, grammar files)
and will send the params to firmware only after the VTSV pipe has been
allocated.

Split the requesting of net/grammar files from sending the parameters to
firmware. The platform driver calls two separate APIs for requesting the files
and for sending the params. The new flow is as follows:

* Enrollment
	- VTSV pipe not allocated
		> Cache the lib addresses
	- VTSV pipe active
		> Cache the lib address
		> Send the new address

* VTSV pipe activated by DAPM
	> Send the lib address

Change-Id: If20ee1f1b7044d831b39d461aa006a2859d2be50
Signed-off-by: Omair Mohammed Abdullah <omair.m.abdullah@intel.com>