intel-mid: Special handling of reboot IPC command
BZ: 195053

Assume reboot syscall happens at the time when a rpmsg is running
and waiting for the IPC completion in kernel. The reboot process
might be scheduled to run. During reboot, the CPUs other than the
CPU running reboot process are shutdown, and local irq on the reboot CPU
is disabled. The previous rpmsg process cannot be scheduled again.
Thus when the reboot rpmsg needs to hold the rpmsg lock, deadlock happened.

One solution is to wait for SCU not busy and send the IPC command without
holding any ipc related locks.

Continue using rpmsg if reboot_force is set, since cpus are not shutdown and
irqs are not disabled when force reboot enabled.

Change-Id: Ia86188607980fcdb5421877d30dff66759c1ed4d
Signed-off-by: Li, Ning <ning.li@intel.com>