EM/THERMAL: Do not update virtual thermal zones during boot
BZ: 173485

The thermal_zone_device_register function inside
thermal_core.c updates the status of the thermal
zone as soon as it is registered. Reading the
temperature of the zone is a part of this update
function.

There are some 'virtual' zones that may depend on
user land for their temperature (because they are
not directly read from sensors, instead calculated
from other sources). For these zones, this patch
disables the update function during boot.

Change-Id: I75e67c8b245b6ba60c71d96c8c7cad303df47697
Signed-off-by: Durgadoss R <durgadoss.r@intel.com>
Signed-off-by: Sumeet Pawnikar <sumeet.r.pawnikar@intel.com>