ov8858: camera sensor doesn't always power on correctly
BZ: 194257

There is a small spike on the ov8858 reset line during the boot,
which causes the ov8858 camera sensor to sometimes fail to power on
correctly.

This was caused by incorrectly setting the default value to high for
the line during the gpio search. Distribute also the delays to ensure
the power lines are stable in all power on cases.

Change-Id: I5adbd21b7d4f09592ac3b7d7fd66f092ccbc4faf
Signed-off-by: Soren Friis <soren.friis@intel.com>