audio:[CHT] Initialize platform clock to OFF
BZ: 195213

Audio Platform clock supplies MCLK for codec.The machine driver
requests this clock to be turned ON and OFF on playing any stream.
But until any stream is played the clock remains ON. Hence request
the clock to be turned OFF in probe.

If the clock is not initialized to OFF, S0ix transitions will be
blocked until any stream is played.

Change-Id: If56fbe5ddf0208e843f22b158cf35fecbfc8d880
Signed-off-by: Jayachandran B <jayachandran.b@intel.com>