dwc3-host-intel: Trigger warm reset to rescue link.
BZ: 197212

On MOFD, plug in some SS device, the link can't train to U0 which will
stay in polling. Trigger warm reset can rescue the link to U0. If not,
the SS device can't be enumeration success.

This patch is one WA for this issue. Will continue debug with silicon
team. After root cause, then revert this patch.

Change-Id: I705b49f20dd80c2b65dea6056d0e29bd81a5f97d
Signed-off-by: Wang, Yu <yu.y.wang@intel.com>