atomisp_platform: Add new flag in atomisp_camera_caps
BZ: 196122

This patch adds a new flag named multi_stream_ctrl
in atomisp_camera_caps. This flag is used only when
the sensor can output multiple streams.
By default this flag is not set. When stream_on call
is invoked in sensor driver, all the output streams
from sensor are started.
If this flag is set, it means that the second stream
also needs to be controlled through stream_on/off calls.

Change-Id: I60f8f1dd1c240134e650e0421ce948db92da843f
Signed-off-by: Tianshu Qiu <tian.shu.qiu@intel.com>