mmc: host: change max_tuning_loop count to be 40
BZ: 199429

Max tuning count is 40. When tuning count reduced to 0, means 40 cycles is
reached. On some pre-silicon environment, it is can reach to 40, which is
maximum, and makes driver thinking the tuning is failed.

Needs to change this value to 41, so that when it reaches to max tuning count,
driver won't think it is a failure case

Change-Id: Icd3e54304e45dd3e78a01e103d4bdafbe8dfca94
Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>