[PORT_FROM_R44B] BYTCR_V2.1/2.2: dw_i2c: Increase PM QoS Req. to correct value
BZ: 202249

It is seen that with the current implementation of holding
PMQOS for C1-1, the cores get stuck at C0 with no load. This
leads to an increased power consumption for LPMP3 usecase.

Change the PM QoS to C6-1. This will allow C1 and that is
safe as the PUNIT will not do Voltage transisions as a part
of cstate change from C0 to C1 through the I2C Interface.

Change-Id: I45b66fd34fadda4b1064e0da75945245999c5c2a
Signed-off-by: Souvik Kumar Chakravarty <souvik.k.chakravarty@intel.com>