drm/i915: unpin sprite when unpin_work is queued
BZ: 208199

When sprite unpin work is being queued to send page flip event, before
execution of the work function if the unpin work is called again then we
end up with an error messge and thereafter the system enters into an
unrecoverable state thereafter. Better to have some work around instead
to making the system unrecoverable. This patch includes that work around.

Change-Id: Ifcc974067591805c7392449fd3f41299cae30b7a
Signed-off-by: Arun R Murthy <arun.r.murthy@intel.com>