PM : Fix for the PANIC on changing the governor dynamically
BZ: 200619

The patch disables touch boost in case the governor is changed
dynamically. For debug purposes, teams use governors,other than Interactive governor by changing the governor through sysfs. The PowerHAL currently is designed for interactive governor and is not aware of such dynamic changes. This was causing a PANIC on a dynamic governor change followed by a touch boost

The patch makes sure that the touch boost is only triggered from PowerHAL,
only when the governor is interactive.

Change-Id: I7c0d5cbd3f4ceccd2e265f895eb927f275757c9c
Signed-off-by: Kumar P Mahesh <mahesh.kumar.p@intel.com>