kernel: Fix flags for stack-protector support
BZ: 204134

The script to check proper implementation of stack-protector by the
compiler does not specify -fno-pic although it is set in KBUILD_CFLAGS.
The compiler should use the -mstack-protector-guard=tls to generate code
suitable for the kernel when this parameter is supported.

Add KBUILD_CFLAGS to the script command.
Add -mstack-protector-guard=tls when supported.

Change-Id: I7627e50450c6faac0bc7cfb0a71a87dd688968c1
Signed-off-by: Laurent FERT <laurent.fert@intel.com>