ASoC: core: Add Fix-up for dai-link
BZ: 201483

In-order to support the DSP's like Intel which needs dynamic configuration
of CODEC<->CODEC dai_link params. For example, BT can be narrow-band or
wide-band. For such instances, a fix-up method is provided to the machine
driver to (re)configure the DAI link params based on the dynamic
selection of backend DAI configuration.

Change-Id: I71b7828317b85e3b82a63b49c890883eb5efa9a0
Signed-off-by: Dharageswari.R <dharageswari.r@intel.com>