drm: add drm_vblank_on interface
BZ: 208756

drm_vblank_off() will turn off vblank interrupts, but as long as the
refcount is elevated drm_vblank_get() will not re-enable them. This
is a problem is someone is holding a vblank reference while drm_vblank_off
is called.

Add drm_vblank_on() as a counterpart to drm_vblank_off() which will
re-enabled vblank interrupts if the refcount is already elevated. This
will allow drivers to choose the specific places in the modeset sequence
at which vblank interrupts get disabled and enabled.

Change-Id: I3f1f272e1284e438f901070e921323379872bbc8
Signed-off-by: Tan Wei <weix.z.tan@intel.com>