SDIO: avoid an error in tuning procedure by ignoring signals
BZ: 207920

Use 'wait_event_timeout' instead of 'wait_event_interruptible_timeout'
when waiting for the interrupt.
There is quite often a signal pending when executing the
tuning procedure, and hence 'wait_event_interruptible_timeout' will
exit prematurely with an error code of -512.
Avoids the:
: Timeout waiting for Buffer Read Ready interrupt during tuning procedure.
Also adds an optimization by not re-reading the interrupt status when
'tuning_done' is set.

Change-Id: I148a4ec9e894e2825451f08d98cc82344fd64673
Signed-off-by: Frode Isaksen <frodex.isaksen@intel.com>