drm/i915: DPI FIFO empty check is not needed
BZ: 216786

This patch is from upstream kernel.
While sending DPI SHUTDOWN command, we cannot wait for FIFO empty as
pipes are not disabled at that time. In case of MIPI we disable port
first and send SHUTDOWN command while pipe is still running and FIFOs
will not be empty, causing spurious error log

Change-Id: If2b565f65c876abefe19fc89454bb09dfe29fa6a
Signed-off-by: Shobhit Kumar <shobhit.kumar@intel.com>
Signed-off-by: Tajunnisha N <tajunnisha.n@intel.com>