[PORT FROM R44B]tlv320aic31xx jack detection
BZ: 210484
Problem:
Google voice is triggered when headphone is inserted/removed.
The root cause: headphone is detected incorrectly as headset and button pressed event is sent,
but not button released event is sent,
a long button pressed event is triggered which is the shortcut key for Google voice app.
Actually, not only headphone is detected incorrectly,
but also headset is detected incorrectly as headphone sometimes.
so the mechanism of jack detection should be improved.

Solution:
1. The comments from TI FAE for stable jack detection
 a. enable BIAS earlier
 b. keep codec debounce time long enough(set to 256ms)
2. like FFRD8 realtek codec implementation. use soc GPIO for jack detection
 a. disable JD interrupt through codec.
 b. byt_codec_hs_detection is used for button press detection only.
 c. Schedule jack insertion/removal detection in soc GPIO interrupt
 d. disable button interrupt in soc GPIO interrupt and enable in byt_enable_hs_button_events
    byt_enable_hs_button_events is delay work which is scheduled when HEADSET inserted

Change-Id: I0a354e44d443f4d44a1cc319798484cc862668f1
Signed-off-by: peter <peter.chen@intel.com>