drm/i915: DRRS: Asynchronously setting the dsi clk
BZ: 217066

In VLV dsi pll dividers register change will impact the PLL only when
we toggle the VCO_EN of the PLL. At present this can be accurately
achieved within Vblank period through SW wait loop only.

But to make sure that userspace calls for the refreshrate change are
not blocked till the vblank, this patch adds an deferred work with
no delay.

Change-Id: I6056904aa8e6154749af3f6779164ff9c76da76b
Signed-off-by: Ramalingam C <ramalingam.c@intel.com>