Audio: mofd: platform: update the mask bit in sst_mix_put
BZ: 211379

Issue : when same register is updated, changing one alsa control
was resulting in wrong update for other alsa control bit resiging
in the same register.
This is because in sst_mix_put soc_dapm_mixer_update_power is called which
inturn calls  soc_widget_update_bits_locked but the mask expected there is
the bit to be updated but the mix_put was always sending 1 as mask
hence it was causing issues, so update the mask accordingly based on the shift.

Change-Id: Iadfe5aa8a251072bc019a0947322e8f1adc30725
Signed-off-by: Mythri P K <mythri.p.k@intel.com>