Fix NULL pointer in SSIC/sysclass interfaces
BZ: 212333

In the SSIC interfaces, pci device is used to caculated the SSIC HCD/xHCI.
When we create the sysclass interfaces, pci sub device are created where HCD
is not the private data of pci sub device. So NULL pointer observed. Use the
SSIC PCI dev to get HCD/xHCI private data can fix this problem.

Change-Id: Ic0aac3bf9bc85f28e262426c8ae2578cd7d5ab93
Signed-off-by: Li Aixiong <aixiong.li@intel.com>